# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI-Powered Research Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /papers/search:
    get:
      operationId: searchPapers
      tags: [papers]
      summary: Search research papers across multiple academic sources
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
          description: Research title, keywords, or problem statement
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
        - name: sources
          in: query
          required: false
          schema:
            type: string
            description: Comma-separated list of sources (semanticscholar,arxiv,openalex,crossref)
      responses:
        "200":
          description: List of relevant research papers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperSearchResult"
        "400":
          description: Bad request
...
  /papers/citations:
    post:
      operationId: generateCitations
      tags: [papers]
      summary: Generate citations in multiple formats
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CitationInput"
      responses:
        "200":
          description: Citations in multiple formats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CitationResult"
  /papers/related:
    post:
      operationId: getRelatedPapers
      tags: [papers]
      summary: Find semantically related papers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RelatedInput"
      responses:
        "200":
          description: List of related papers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperSearchResult"
...
  /searches/history:
    get:
      operationId: getSearchHistory
      tags: [papers]
      summary: Get recent search history for personalized recommendations
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Search history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchHistoryResult"
  /bookmarks:
    get:
      operationId: getBookmarks
      tags: [papers]
      summary: Get saved/bookmarked papers
      responses:
        "200":
          description: List of bookmarked papers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookmarkListResult"
    post:
      operationId: addBookmark
      tags: [papers]
      summary: Save a paper to bookmarks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookmarkInput"
      responses:
        "201":
          description: Bookmarked paper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
  /bookmarks/{paperId}:
    delete:
      operationId: removeBookmark
      tags: [papers]
      summary: Remove a paper from bookmarks
      parameters:
        - name: paperId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bookmark removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Paper:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        year:
          type: ["integer", "null"]
        journal:
          type: ["string", "null"]
        abstract:
          type: ["string", "null"]
        doi:
          type: ["string", "null"]
        citationCount:
          type: ["integer", "null"]
        pdfUrl:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        source:
          type: string
        hasDataset:
          type: boolean
        githubUrl:
          type: ["string", "null"]
        difficultyLevel:
          type: ["string", "null"]
          description: "beginner | intermediate | advanced"
        qualityScore:
          type: ["number", "null"]
        isOpenAccess:
          type: boolean
        fields:
          type: array
          items:
            type: string
      required:
        - id
        - title
        - authors
        - source
        - isOpenAccess
        - hasDataset
        - fields
    PaperSearchResult:
      type: object
      properties:
        papers:
          type: array
          items:
            $ref: "#/components/schemas/Paper"
        total:
          type: integer
        query:
          type: string
      required:
        - papers
        - total
        - query
...
    CitationInput:
      type: object
      properties:
        paper:
          $ref: "#/components/schemas/Paper"
      required:
        - paper
    CitationResult:
      type: object
      properties:
        apa:
          type: string
        ieee:
          type: string
        mla:
          type: string
        chicago:
          type: string
        bibtex:
          type: string
      required:
        - apa
        - ieee
        - mla
        - chicago
        - bibtex
    RelatedInput:
      type: object
      properties:
        paper:
          $ref: "#/components/schemas/Paper"
        limit:
          type: integer
          default: 5
      required:
        - paper
...
    SearchHistory:
      type: object
      properties:
        id:
          type: integer
        query:
          type: string
        resultCount:
          type: integer
        searchedAt:
          type: string
      required:
        - id
        - query
        - resultCount
        - searchedAt
    SearchHistoryResult:
      type: object
      properties:
        history:
          type: array
          items:
            $ref: "#/components/schemas/SearchHistory"
      required:
        - history
    Bookmark:
      type: object
      properties:
        id:
          type: integer
        paperId:
          type: string
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        year:
          type: ["integer", "null"]
        url:
          type: ["string", "null"]
        source:
          type: string
        savedAt:
          type: string
      required:
        - id
        - paperId
        - title
        - authors
        - source
        - savedAt
    BookmarkListResult:
      type: object
      properties:
        bookmarks:
          type: array
          items:
            $ref: "#/components/schemas/Bookmark"
      required:
        - bookmarks
    BookmarkInput:
      type: object
      properties:
        paperId:
          type: string
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        year:
          type: ["integer", "null"]
        url:
          type: ["string", "null"]
        source:
          type: string
      required:
        - paperId
        - title
        - authors
        - source