# lib/api-zod/src/generated/types/bookmark.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Research Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface Bookmark {
  id: number;
  paperId: string;
  title: string;
  authors: string[];
  /** @nullable */
  year?: number | null;
  /** @nullable */
  url?: string | null;
  source: string;
  savedAt: string;
}