# lib/api-zod/src/generated/types/citationResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Research Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface CitationResult {
  apa: string;
  ieee: string;
  mla: string;
  chicago: string;
  bibtex: string;
}