# lib/api-zod/src/generated/types/relatedInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Research Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { Paper } from './paper';

export interface RelatedInput {
  paper: Paper;
  limit?: number;
}